skills-verify/pyproject.toml
[project]
name = "skills-verify"
version = "0.1.0"
description = "Verification harness for generated opencli skills"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents",
    "langchain-core",
    "langchain-openai",
    "python-dotenv",
]

[tool.uv]
package = false
